[K8S] Refactored failure hooks data structure, added check_only option
check_only will only check if the resource is declared on the cluster.
It will stop immediately if it has not. This is mainly for things such
as secrets or persistent volumes. The idea is to have the converge
check if they are present, but get human intervention if they are not
available.